Moved implementation of gdt to cpp file
